+++ backend/app/controller/admin/container_log.go
@@ -4,6 +4,7 @@ import (
 	"rustdesk-api-server-pro/app/model"
 	"rustdesk-api-server-pro/config"
 	"rustdesk-api-server-pro/db"
+	"rustdesk-api-server-pro/internal/errcode"
 
 	"github.com/kataras/iris/v12"
 	"github.com/kataras/iris/v12/mvc"
@@ -100,6 +101,10 @@ func (c *ContainerLogController) HandleList() mvc.Result {
 }
 
 func (c *ContainerLogController) HandleClear() mvc.Result {
+	user := c.GetUser()
+	if user == nil || !user.IsAdmin {
+		return c.Error(nil, errcode.New(errcode.ERR1010.Code, errcode.ERR1010.Message).Error())
+	}
 	_, err := c.Db.Exec("DELETE FROM container_log")
 	if err != nil {
 		return c.dbError(err)
